Add a new hash class in __gnu_ext for the extension containers. There are two motivations for this.
First, this allows users who are specializing __gnu_ext::hash to
continue doing so without changing their code.

Second, SGI specifies hash overloads for char* and const char* that
perform a hash of the string, not of the pointer.

In order to support this, the hashing code for string is factored out.

git-svn-id: svn://svn.chromium.org/llvm-project/libcxx/trunk@136539 0b72dbe1-c17f-4bc7-b9db-2b4152be0356