[libc++] random_device fails if open returns zero
random_device::random_device(const string&) wrongly assumes that open
can only validly return a file descriptor greater than zero.

This results in random_device believing that it didn't successfully open
the device causing it to throw in it's constructor, this ends up leaking
a file descriptor.

The fix is simple, don't error on file descriptors which are zero.


git-svn-id: svn://svn.chromium.org/llvm-project/libcxx/trunk@210060 0b72dbe1-c17f-4bc7-b9db-2b4152be0356