Don't hardcode the Czech locale name.
We already have a definition for the Czech locale name in
platform_support.h. Use this one instead.

While there, respect the common format of the tests. For most other
tests it's the case that test_iterators.h is placed right underneath the
other #includes (without an empty line). platform_support.h is included
after an empty line.


git-svn-id: svn://svn.chromium.org/llvm-project/libcxx/trunk@232383 0b72dbe1-c17f-4bc7-b9db-2b4152be0356