Fix race conditions in test class used throughout the std::thread tests.
The test class 'G' reads and writes to the same static variables in its
constructor, destructor and call operator. When threads are
constructed using `std::thread t((G()))` there is a race condition between the
destruction of the temporary and the execution of `G::operator()()`.

The fix is to simply create the input before creating the thread.


git-svn-id: svn://svn.chromium.org/llvm-project/libcxx/trunk@233946 0b72dbe1-c17f-4bc7-b9db-2b4152be0356