re.results.form: Format out-of-range subexpression references as null
Rather than crashing in match_results::format() when a reference to a
marked subexpression is out of range, format the subexpression as empty
(i.e., replace it with an empty string).  Note that
match_results::operator[]() has a range-check and returns a null match
in this case, so this just re-uses that logic.

git-svn-id: svn://svn.chromium.org/llvm-project/libcxx/trunk@259682 0b72dbe1-c17f-4bc7-b9db-2b4152be0356